Separate maxmind from sort
